Update generated code to 6637c8
Changes to `mix_deps_get_and_compile/0` are to fix the following error

    1) test inspect/2 does not include password (RainyDay.AccountsTest)
      apps/rainy_day/test/rainy_day/accounts_test.exs:477
      Refute with =~ failed
      code:  refute inspect(%User{password: "123456"}) =~ "password: \"123456\""
      left:  "%RainyDay.Accounts.User{__meta__: #Ecto.Schema.Metadata<:built, \"users\">, confirmed_at: nil, email: nil, hashed_password: nil, id: nil, inserted_at: nil, password: \"123456\", updated_at: nil}"
      right: "password: \"123456\""
      stacktrace:
        test/rainy_day/accounts_test.exs:478: (test)

I think the problem was the Inspect protocol was compiled for the dependencies,
but not recompiled with the updated application code. By running `mix compile`
instead of `mix deps.compile` it seems to correctly recompile the inspect
protocol so the tests pass.

Also, since Elixir < 1.8, does not support deriving the Inspect protocol (with
the :except option), this conditionally adds a custom Inspect implementation
when the project is generated on Elixir 1.7.