maybe_store_return_to handles query string
Before this change, maybe_store_return_to stores the `request_path`
attribute of the Plug.Conn, but does not handle the `query_string` if it
is present.

Now, we append the `query_string` to the `request_path` if it is
present, in order to paths with query strings, and without them.